requirements.txt
neo4j>=5.0.0
pandas>=2.0.0
openai>=1.0.0
python-dotenv>=1.0.0
rich>=13.0.0
streamlit>=1.32.0
